Handle videos that lack durations
Videos that haven't premiered yet still show up in the all-uploads playlist for
the channel. However, the API doesn't give the duration info for that video,
which was causing the video list tool to fail.

Now I exclude videos that lack durations from the list, on the grounds that
they're unwatchable anyway, so they shouldn't be in the list.