Handle videos with titles in multiple languages.
YouTube allows video titles to be in multiple languages. The primary
title of the video is in the `title' field of the `snippet', while
localized titles are present in a `localizations' object.

This change makes it so youtube-video-list prefers the English
localization if it is available, and falls back to the primary video
title if it is not.